Voor wijziging aparte klasse Scheduler

+++ APITaskManagement.Logic/Api/Api.cs
@@ -41,6 +41,7 @@ namespace APITaskManagement.Logic.Api
 
         public void SendRequestsToTarget(Common.HttpMethod httpMethod, Url url, Authentication authentication, Task task)
         {
+            
             using (HttpClient client = new HttpClient())
             {
                 HttpResponseMessage responseMessage = new HttpResponseMessage(System.Net.HttpStatusCode.OK);
@@ -71,44 +72,53 @@ namespace APITaskManagement.Logic.Api
 
                 Requests = GetRequestsForTask(task.Id);
 
+                HashSet<int> keys = new HashSet<int>();
                 foreach (Request request in Requests)
                 {
                     // Check if Url is reachable
                     if (HostIsReachable(url.Address))
                     {
-
-                        switch (httpMethod)
+                        if (!keys.Contains(request.ReferenceId))
                         {
-                            case Common.HttpMethod.Get:
-                                responseMessage = client.GetAsync(url.Address).Result;
-                                break;
-                            case Common.HttpMethod.Post:
-                                responseMessage = client.PostAsync(url.Address, new StringContent(request.Body, Encoding.UTF8, "application/json")).Result;
-                                break;
-                            case Common.HttpMethod.Put:
-                                responseMessage = client.PutAsync(url.Address, new StringContent(request.Body, Encoding.UTF8, "application/json")).Result;
-                                break;
-                            case Common.HttpMethod.Patch:
-                                responseMessage = client.PutAsync(url.Address, new StringContent(request.Body, Encoding.UTF8, "application/json")).Result;
-                                break;
-                            case Common.HttpMethod.Delete:
-                                responseMessage = client.DeleteAsync(url.Address).Result;
-                                break;
-                            default:
-                                responseMessage = client.GetAsync(url.Address).Result;
-                                break;
+                            switch (httpMethod)
+                            {
+                                case Common.HttpMethod.Get:
+                                    responseMessage = client.GetAsync(url.Address).Result;
+                                    break;
+                                case Common.HttpMethod.Post:
+                                    responseMessage = client.PostAsync(url.Address, new StringContent(request.Body, Encoding.UTF8, "application/json")).Result;
+                                   
+                                    break;
+                                case Common.HttpMethod.Put:
+                                    responseMessage = client.PutAsync(url.Address, new StringContent(request.Body, Encoding.UTF8, "application/json")).Result;
+                                    break;
+                                case Common.HttpMethod.Patch:
+                                    responseMessage = client.PutAsync(url.Address, new StringContent(request.Body, Encoding.UTF8, "application/json")).Result;
+                                    break;
+                                case Common.HttpMethod.Delete:
+                                    responseMessage = client.DeleteAsync(url.Address).Result;
+                                    break;
+                                default:
+                                    responseMessage = client.GetAsync(url.Address).Result;
+                                    break;
+                            }
+
+                            var result = responseMessage.Content.ReadAsStringAsync().Result;
+                            var statusCode = (int)responseMessage.StatusCode;
+                            var description = responseMessage.StatusCode.ToString();
+
+                            var response = new Response(statusCode, description, result);
+                            request.SetResponse(response);
+
+                            if (request.ExecPost == true)
+                            {
+                                ExecutePost(request);
+                            }
+
+                            LogResponse(request, url, task.SPLogger);
+
+                            keys.Add(request.ReferenceId);
                         }
-
-                        var result = responseMessage.Content.ReadAsStringAsync().Result;
-                        var statusCode = (int)responseMessage.StatusCode;
-                        var description = responseMessage.StatusCode.ToString();
-
-                        var response = new Response(statusCode, description, result);
-                        request.SetResponse(response);
-
-                        LogResponse(request, url, task.SPLogger);
-
-                        
                     }
                     else
                     {
@@ -152,6 +162,8 @@ namespace APITaskManagement.Logic.Api
         }
 
         protected abstract IList<Request> GetRequestsForTask(Guid taskId);
+
+        protected abstract void ExecutePost(Request request);
 
         public int GetNumberOfRequests()
         {
